Rename folders to lowercase for conformance.
