+++ pyisy/connection.py
