+++ pyisy/constants.py
