+++ pyisy/isy.py
@@ -17,10 +17,10 @@ from .constants import (
 )
 from .events import EventStream
 from .Nodes import Nodes
-from .Nodes.node import Node
+from .nodes.node import Node
 from .Programs import Programs
-from .Programs.folder import Folder
-from .Programs.program import Program
+from .programs.folder import Folder
+from .programs.program import Program
 from .Variables import Variables
 
 
@@ -49,7 +49,7 @@ class ISY:
     :ivar connected: Read only boolean value indicating if the class is
                      connected to the controller.
     :ivar log: Logger used by the class and its children.
-    :ivar nodes: :class:`~PyISY.Nodes.Nodes` manager that interacts with
+    :ivar nodes: :class:`~PyISY.nodes.Nodes` manager that interacts with
                  Insteon nodes and groups.
     :ivar programs: Program manager that interacts with ISY programs and i
                     folders.
