+++ pyisy/nodes/group.py
