+++ pyisy/nodes/node.py
