+++ pyisy/nodes/nodebase.py
