+++ pyisy/strings.py
