+++ pyisy/variables/variable.py
